+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.22 (unreleased)
 
 * Remove hard dependency on `pkg` subfolder in app project folder
+* Report error when `install.sh` is run not from `build` folder
 
 ## 0.0.21 (2024-03-26)